fix swagger codegen integration for depot client
The resttemplate library (declared in pom.xml under swagger codegen plugin), annotates generated classes as spring components but configured values are not annotated. This makes it difficult to further customize these components. There is no way to tell codegen to turn off annotations and so introducing a configuration that produces customized components results in multiple bean definitions.

This fix uses the resteasy library which does not annotate generated components. As such, Depotconfiguration can provide producers which are correctly configured.

Also:
 - upgraded spring boot parent to the latest version
 - downgraded swagger codegen plugin from 3.0.22 to 3.0.20 which is the latest version which works without breaking spring logging (see pom.xml properties for comment)